Introduce `Stringifier`
Also:
* Renamed `ConverterRegistry` to `StringifierRegistry`.
* Moved stringification tests into separate test suite.
